Introduce d3 v4 micro modules. Move utilities to utils directory
