+++ package.json
@@ -20,13 +20,16 @@
     "autobind-decorator": "^1.3.3",
     "bowser": "^1.0.0",
     "canvas-composite-types": "^1.0.4",
-    "d3": "^3.5.16",
+    "d3-array": "^1.0.1",
+    "d3-color": "^1.0.1",
+    "d3-format": "^1.0.2",
+    "d3-geo": "^1.2.4",
+    "d3-scale": "^1.0.3",
+    "d3-selection": "^1.0.2",
     "global": "^4.3.0",
-    "keymirror": "^0.1.1",
     "mapbox-gl": "0.21.0",
     "pure-render-decorator": "^1.1.0",
     "svg-transform": "0.0.3",
-    "tap-browser-color": "^0.1.2",
     "viewport-mercator-project": "^2.0.1"
   },
   "devDependencies": {
@@ -40,6 +43,8 @@
     "babelify": "^7.2.0",
     "browserify": "^13.0.0",
     "budo": "^8.2.2",
+    "d3-random": "^1.0.1",
+    "disc": "^1.3.2",
     "electron-prebuilt": "^0.37.3",
     "envify": "^3.4.0",
     "eslint": "^1.10.3",
@@ -49,6 +54,7 @@
     "react-dom": "15.x",
     "tape": "^4.5.1",
     "tape-catch": "^1.0.4",
+    "tap-browser-color": "^0.1.2",
     "testron": "^1.2.0",
     "uber-standard": "^5.1.0"
   },
@@ -59,7 +65,8 @@
     "test": "npm run lint && browserify -t babelify -t envify test/index.js | testron",
     "precommit": "npm run lint -s",
     "prepush": "npm run lint -s",
-    "prepublish": "npm run build"
+    "prepublish": "npm run build",
+    "profile-disc": "browserify src/index.js --full-paths -t babelify | discify --open"
   },
   "files": [
     "src",
